+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-7646](https://openlmis.atlassian.net/browse/OLMIS-7646): Add functionality to generate programOrderable.csv
 * [OLMIS-7645](https://openlmis.atlassian.net/browse/OLMIS-7645): Add functionality to generate orderableIdentifier.csv
 * [OLMIS-7647](https://openlmis.atlassian.net/browse/OLMIS-7647): Add functionality to generate tradeItem.csv
+* [OLMIS-7648](https://openlmis.atlassian.net/browse/OLMIS-7648): Add mapping files
 
 15.2.4 / 2022-10-07
 ==================